% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviation.R
\name{abbr_journal_name}
\alias{abbr_journal_name}
\title{Journal name abbreviation}
\usage{
abbr_journal_name(x)
}
\arguments{
\item{x}{journal name}
}
\description{
Journal name abbreviation
}
\examples{
abbr_journal_name("Proceedings of the National Academy of Sciences")
}
