% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{article_info}
\alias{article_info}
\title{Update article information}
\usage{
article_info(x, user, repo, number, labels = NULL, ...)
}
\arguments{
\item{x}{x}

\item{user}{user account name}

\item{repo}{repository}

\item{number}{issue number}

\item{labels}{issue labels}
}
\description{
Update article information
}
