% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_template.R
\name{create_issue_template}
\alias{create_issue_template}
\title{Create issue template}
\usage{
create_issue_template(user, repo)
}
\arguments{
\item{user}{user account name}

\item{repo}{repository}
}
\description{
Create issue template
}
\examples{
\dontrun{
create_issue_template("uribo", "bucky_sandbox")
}
}
