% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{make_issue_body}
\alias{make_issue_body}
\title{Make GitHub issue's body contents}
\usage{
make_issue_body(x, style = "oikos", ...)
}
\arguments{
\item{x}{GitHub issue title}

\item{style}{a CSL style (for text format only). See \code{\link[rcrossref:get_styles]{get_styles()}}
for options. Default: 'apa'. If there's a style that CrossRef doesn't support
you'll get a  \verb{(500) Internal Server Error}}

\item{...}{path to \link[rcrossref:cr_cn]{rcrossref::cr_cn}}
}
\description{
Make GitHub issue's body contents
}
\examples{
\dontrun{
make_issue_body("DOI: 10.1016/j.tourman.2019.104010")
}
}
